+++ src/main/java/org/kumoricon/registration/model/attendee/AttendeeFactory.java
@@ -18,6 +18,7 @@ public class AttendeeFactory {
         Attendee attendee = new Attendee();
         attendee.setFirstName("Firstname");
         attendee.setLastName("Lastname");
+        attendee.setPreferredPronoun("He/Him/His");
         attendee.setFanName("Fan Name - Adult");
         attendee.setBadgeNumber("TST12340");
         attendee.setBadge(badge);