+++ CHANGES.md
@@ -6,6 +6,7 @@ Version 0.6.0 (Pending)
 - Added Force Password Change page/interceptor. This included a database schema change, making it incompatible
   with the previous version.
 - Log current user with every message
+- Added preferred pronoun to Attendee class and database table (schema change)
 
 Version 0.5 (4/9/2019)
 - Initial "release" version. Not ready for production yet, but start tracking changes